+++ packages/antd-management-fast-design-playground/src/components/displayer/DocumentPrintDesigner/DocumentContent/Cells/CellMoney/index.jsx
@@ -91,7 +91,8 @@ class CellMoney extends CellBase {
               <div
                 style={{
                   ...frontSizeStyle,
-                  fontSize: '16px',
+                  fontSize: '20px',
+                  fontWeight: 'bold',
                   ...fontFamilyStyle,
                   ...colorStyle,
                 }}