+++ examples/example-antd-management-fast-framework/CHANGELOG.md
@@ -1,5 +1,19 @@
 # Change Log
 
+## 1.9.83
+
+### Patch Changes
+
+- update
+- Updated dependencies
+  - antd-management-fast-design-playground@1.0.206
+  - antd-management-fast-design-react@1.0.133
+  - antd-management-fast-component@2.7.51
+  - antd-management-fast-framework@2.12.57
+  - antd-management-fast-formily@1.0.146
+  - antd-management-fast-common@2.6.164
+  - antd-management-fast-flow@1.0.134
+
 ## 1.9.82
 
 ### Patch Changes