+++ packages/antd-management-fast-formily/CHANGELOG.md
@@ -1,5 +1,15 @@
 # antd-management-fast-formily
 
+## 1.0.146
+
+### Patch Changes
+
+- update
+- Updated dependencies
+  - antd-management-fast-design-react@1.0.133
+  - antd-management-fast-component@2.7.51
+  - antd-management-fast-common@2.6.164
+
 ## 1.0.145
 
 ### Patch Changes